requirements.txt
--extra-index-url=https://pypi.viewflow.io/A435DFF5AB77/simple/
boto3==1.9.39
botocore==1.12.39
dj-database-url==0.5.0
Django==2.1.2
django-autofixture==0.12.1
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-formtools==2.1
django-material==1.4.1
django-material-pro==1.4.1
django-storages==1.7.1
django-superform==0.3.1
django-template-debug==0.3.5
django-viewflow==1.4.0
django-viewflow-pro==1.4.1
docutils==0.14
filelock==3.0.9
gunicorn==19.9.0
jmespath==0.9.3
pluggy==0.8.0
py==1.7.0
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.7
s3transfer==0.1.13
six==1.11.0
sqlparse==0.2.4
toml==0.10.0
tox==3.5.3
urllib3==1.24.1
virtualenv==16.1.0
